+++ Source/OxyPlot.Wpf.Tests/PngAssert.cs
@@ -36,17 +36,21 @@ namespace OxyPlot.Wpf.Tests
             {
                 EnsureFolder(expected);
                 File.Copy(actual, expected);
-                Assert.Fail("File not found: {0}", expected);
+                Assert.Fail($"File not found: {expected}");
             }
 
             if (expectedImage.GetLength(0) != actualImage.GetLength(0))
             {
-                Assert.Fail("Expected height: {0}\nActual height:{1}\n{2}", expectedImage.GetLength(0), actualImage.GetLength(0), message);
+                Assert.Fail($@"Expected height: {expectedImage.GetLength(0)}
+Actual height: {actualImage.GetLength(0)}
+{message}");
             }
 
             if (expectedImage.GetLength(1) != actualImage.GetLength(1))
             {
-                Assert.Fail("Expected width: {0}\nActual width:{1}\n{2}", expectedImage.GetLength(1), actualImage.GetLength(1), message);
+                Assert.Fail($@"Expected width: {expectedImage.GetLength(1)}
+Actual width: {actualImage.GetLength(1)}
+{message}");
             }
 
             var w = expectedImage.GetLength(0);
@@ -88,13 +92,11 @@ namespace OxyPlot.Wpf.Tests
                     File.WriteAllBytes(output, encoder.Encode(differenceImage));
                 }
 
-                Assert.Fail(
-                    "{0}:\nPixel differences: {1}\nExpected image: {2}\nActual image: {3}\nDiff image: {4}",
-                    message,
-                    differences,
-                    Path.GetFullPath(expected),
-                    Path.GetFullPath(actual),
-                    Path.GetFullPath(output));
+                Assert.Fail($@"{message}:
+Pixel differences: {differences}
+Expected image: {Path.GetFullPath(expected)}
+Actual image: {Path.GetFullPath(actual)}
+Diff image: {Path.GetFullPath(output)}");
             }
         }